[feature] Stripe: convert cents for real

+++ src/main/java/com/orama/e_commerce/dtos/stripe_entities/ProductRequest.java
@@ -1,3 +1,5 @@
 package com.orama.e_commerce.dtos.stripe_entities;
 
-public record ProductRequest(Long amount, Long quantity, String name, String currency) {}
+import java.math.BigDecimal;
+
+public record ProductRequest(BigDecimal amount, Long quantity, String name, String currency) {}
